# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Это пример сервера с использованием Swagger в Gin.",
        "title": "swagger API",
        "contact": {},
        "version": "1.0"
    },
    "host": "157.90.227.125:8000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/contest": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "получить активные контесты",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "contest_id",
                        "name": "contest_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-array_schema_ContestData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    }
                }
            }
        },
        "/api/v1/contest/active": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "получить активные контесты",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_ContestActivity"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    }
                }
            }
        },
        "/api/v1/contest/book": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "получить книги контеста",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "contest_id",
                        "name": "contest_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-array_model_ContestBook"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    }
                }
            }
        },
        "/api/v1/contest/join": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "подключиться к розыгрышу",
                "parameters": [
                    {
                        "description": "CreateMark",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.JoinContestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    }
                }
            }
        },
        "/api/v1/contest/prize": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "получить призы контеста",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "contest_id",
                        "name": "contest_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-array_model_ContestPrize"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    }
                }
            }
        },
        "/api/v1/contest/read": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contest"
                ],
                "summary": "Прочесть контест книгу",
                "parameters": [
                    {
                        "description": "CreateMark",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.ReadContestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_ContestPassBook"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.ContestBook": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "contest_book_id": {
                    "type": "integer"
                },
                "contest_coins": {
                    "type": "integer"
                },
                "contest_history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ContestHistory"
                    }
                },
                "contest_id": {
                    "type": "integer"
                },
                "count_of_questions": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "day_number": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "photo_path": {
                    "type": "string"
                },
                "point": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.ContestHistory": {
            "type": "object",
            "properties": {
                "contest_book_id": {
                    "type": "integer"
                },
                "contest_history_id": {
                    "type": "integer"
                },
                "contest_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "points": {
                    "type": "integer"
                },
                "read_time": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "model.ContestPrize": {
            "type": "object",
            "properties": {
                "contest_id": {
                    "type": "integer"
                },
                "contest_prize_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "photo_path": {
                    "type": "string"
                },
                "prize_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "schema.ContestActivity": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.ContestData"
                    }
                },
                "ended": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.ContestResultPrize"
                    }
                }
            }
        },
        "schema.ContestData": {
            "type": "object",
            "properties": {
                "already_joined": {
                    "type": "boolean"
                },
                "contest_id": {
                    "type": "integer"
                },
                "current_day_number": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "total_users_count": {
                    "type": "integer"
                }
            }
        },
        "schema.ContestPassBook": {
            "type": "object",
            "properties": {
                "coins": {
                    "type": "integer"
                },
                "points": {
                    "type": "integer"
                }
            }
        },
        "schema.ContestResultPrize": {
            "type": "object",
            "properties": {
                "consolation_prize_sapphire": {
                    "type": "integer"
                },
                "contest_id": {
                    "type": "integer"
                },
                "number": {
                    "type": "integer"
                },
                "prize": {
                    "$ref": "#/definitions/model.ContestPrize"
                }
            }
        },
...
        "schema.Empty": {
            "type": "object"
        },
        "schema.JoinContestRequest": {
            "type": "object",
            "properties": {
                "contest_id": {
                    "type": "integer"
                }
            }
        },
...
        "schema.ReadContestRequest": {
            "type": "object",
            "properties": {
                "contest_book_id": {
                    "type": "integer"
                },
                "read_time": {
                    "type": "integer"
                },
                "right_questions": {
                    "type": "integer"
                }
            }
        },
...
        "schema.Response-array_model_ContestBook": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ContestBook"
                    }
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "schema.Response-array_model_ContestPrize": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ContestPrize"
                    }
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "schema.Response-array_schema_ContestData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.ContestData"
                    }
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "schema.Response-schema_ContestActivity": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/schema.ContestActivity"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "schema.Response-schema_ContestPassBook": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/schema.ContestPassBook"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "schema.Response-schema_Empty": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/schema.Empty"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}